Fix sim cacheing
