+++ packages/core/src/sims/processors/sim_processors.ts
@@ -1,7 +1,7 @@
 import { JOB_DATA, JobName } from "@xivgear/xivmath/xivconstants";
 import { CycleSettings } from "@xivgear/core/sims/cycle_settings";
 import { CharacterGearSet } from "@xivgear/core/gear";
-import { sum } from "@xivgear/core/util/array_utils";
+import { arrayEq, sum } from "@xivgear/core/util/array_utils";
 import { addValues, applyStdDev, multiplyFixed } from "@xivgear/xivmath/deviation";
 import { PartyBuff, SimSettings, SimSpec, Simulation } from "@xivgear/core/sims/sim_types";
 import {
@@ -17,6 +17,8 @@ import {
     Rotation
 } from "@xivgear/core/sims/cycle_sim";
 import { BuffSettingsManager } from "@xivgear/core/sims/common/party_comp_settings";
+
+export type RotationCacheKey = (number | boolean | string)[];
 
 /**
  * Base class for a CycleProcessor based simulation. You should extend this class,
@@ -50,7 +52,8 @@ implements Simulation<FullResultType, InternalSettingsType, ExternalCycleSetting
     readonly manualRun = false;
 
     private cachedCycleProcessors: [string, CycleProcessor][];
-    private cachedSpeed: number;
+    private cachedRotationKey: RotationCacheKey;
+
     protected constructor(public readonly job: JobName, settings?: ExternalCycleSettings<InternalSettingsType>) {
         this.settings = this.makeDefaultSettings();
         if (settings !== undefined) {
@@ -188,12 +191,20 @@ implements Simulation<FullResultType, InternalSettingsType, ExternalCycleSetting
         };
     }
 
+    protected computeCacheKey(set: CharacterGearSet): RotationCacheKey {
+        const stats = set.computedStats;
+        const sps = stats.spellspeed;
+        const sks = stats.skillspeed;
+        const wdly = stats.weaponDelay;
+        return [sps, sks, wdly];
+    }
+
     async simulate(set: CharacterGearSet): Promise<FullResultType> {
         console.debug("Sim start");
-        const setSpeed = set.isStatRelevant('spellspeed') ? set.computedStats.spellspeed : set.computedStats.skillspeed;
-        if (setSpeed !== this.cachedSpeed) {
+        const cacheKey = this.computeCacheKey(set);
+        if (!arrayEq(this.cachedRotationKey, cacheKey)) {
             this.cachedCycleProcessors = this.generateRotations(set);
-            this.cachedSpeed = setSpeed;
+            this.cachedRotationKey = cacheKey;
         }
         return this.calcDamage(set);
     };
