+++ packages/math-frontend/src/scripts/nav_hash.ts
@@ -2,33 +2,11 @@ import {CALC_HASH, HASH_QUERY_PARAM, PATH_SEPARATOR, splitPath} from "@xivgear/c
 
 import {formatTopMenu} from "./base_ui";
 import {openMath} from "./mathpage/math_ui";
+import {arrayEq} from "@xivgear/core/util/array_utils";
 
 let expectedHash: string[] | undefined = undefined;
 
 
-/**
- * Determine if two arrays have equal members. Not a deep equals - only inspects one level.
- *
- * @param left The first array
- * @param right The second array
- */
-export function arrayEq(left: unknown[] | undefined, right: unknown[] | undefined) {
-    if (left === undefined && right === undefined) {
-        return true;
-    }
-    if (left === undefined || right === undefined) {
-        return false;
-    }
-    if (left.length !== right.length) {
-        return false;
-    }
-    for (let i = 0; i < left.length; i++) {
-        if (left[i] !== right[i]) {
-            return false;
-        }
-    }
-    return true;
-}
 
 /**
  * Get the current page path
